Change weight type from double to int

+++ mtsp-vrp-c/include/mtsp-vrp-c.h
@@ -15,7 +15,7 @@
 
 extern "C"
 {
-    MTSP_VRP_C_EXPORT int solve_mtsp_vrp(size_t numberOfAgents, size_t numberOfNodes, const int* start_positions, const int* end_positions, const double* weights, int timeout,
+    MTSP_VRP_C_EXPORT int solve_mtsp_vrp(size_t numberOfAgents, size_t numberOfNodes, const int* start_positions, const int* end_positions, const int* weights, int timeout,
         double* lowerBound, double* upperBound, int* paths, size_t* pathOffsets);
 }
